Moved docs around and updated tests
