+++ src/components/address-finder/autosuggest.address.dom.js
@@ -1,14 +1,7 @@
 import domready from 'js/domready';
 
-async function initialiseAutosuggests() {
-  const autosuggests = [...document.querySelectorAll('.js-autosuggest')];
+async function initialiseAddressAutosuggests() {
   const addressAutosuggests = [...document.querySelectorAll('.js-address-autosuggest')];
-
-  if (autosuggests.length) {
-    const Autosuggest = (await import('./autosuggest')).default;
-
-    autosuggests.forEach(autosuggest => new Autosuggest(autosuggest));
-  }
 
   if (addressAutosuggests.length) {
     const AutosuggestAddress = (await import('./autosuggest.address')).default;
@@ -17,4 +10,4 @@ async function initialiseAutosuggests() {
   }
 }
 
-domready(initialiseAutosuggests);
+domready(initialiseAddressAutosuggests);
