+++ src/components/address-finder/examples/address-finder-editable/index.njk
@@ -7,10 +7,8 @@
         "text": "Enter address or postcode and select from results",
         "id": "address-label"
     },
-    "isEditable": false,
+    "isEditable": true,
     "mandatory": true,
-    "externalInitialiser": true,
-    "autocomplete": "new-password",
     "APIDomain": "https://whitelodge-eq-ai-api.census-gcp.onsdigital.uk",
     'APIDomainBearerToken': "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.5W2EgODHpsts672J4psiGH31H-vJBz1r56_Tr9_8pPU",
     "instructions": "Use up and down keys to navigate suggestions once you\'ve typed more than two characters. Use the enter key to select a suggestion. Touch device users, explore by touch or with swipe gestures.",
@@ -27,14 +25,28 @@
     "noResults": "No results found. Try entering a different part of the address",
     "tooManyResults": "{n} results found. Enter more of the address to improve results",
     "typeMore": "Enter more of the address to get results",
+    "autocomplete": "new-password",
     "errorTitle": "There is a problem with your answer",
     "errorMessageEnter": "Enter an address",
     "errorMessageSelect": "Select an address",
     "errorMessageAPI": "Sorry, there was a problem loading addresses",
-    "errorMessageAPILinkText": "Go back to homepage",
-    "errorMessageAPILink": "#",
+    "errorMessageAPILinkText": "Enter address manually",
     "options": {
         "regionCode": "gb-eng",
         "addressType": "residential"
-    }
+    },
+    "line1": {
+        "label": "Address line 1"
+    },
+    "line2": {
+        "label": "Address line 2"
+    },
+    "town": {
+        "label": "Town or city"
+    },
+    "postcode": {
+        "label": "Postcode"
+    },
+    "searchButton": "Search for an address",
+    "manualLinkText": "Manually enter address"
 }) }}
