+++ src/components/autosuggest/_test-template.njk
@@ -1,0 +1,3 @@
+{% from "components/address-finder/_macro.njk" import onsAddressFinder %}
+
+{{ onsAddressFinder(params) }}
