+++ src/components/autosuggest/index.njk
@@ -9,14 +9,11 @@ Autosuggests are used to suggest options to users as they type. The user can cho
 
 You can search topics such as countries, languages, nationalities and job types or you can create your own.
 
-### How to use
-
-Autosuggest functionality can be created using the [input](/components/input) component and adding the necessary parameters inside the `autosuggest` key.
-Look at the Nunjucks output on the example below to see how it should be configured. 
-
 {{
     patternlibExample({"path": "components/autosuggest/examples/autosuggest-country/index.njk"})
 }}
+
+### How to use
 
 The list needs to be made available as a `.json` file. To query the list we use a JavaScript library called {{
         onsExternalLink({
