+++ src/js/index.js
@@ -13,7 +13,8 @@ import 'components/uac/uac.dom';
 import 'components/relationships/relationships.dom';
 import 'components/checkboxes/checkboxes.dom';
 import 'components/radios/radios.dom';
-import 'components/input/autosuggest/autosuggest.dom';
+import 'components/autosuggest/autosuggest.dom';
+import 'components/address-finder/autosuggest.address.dom';
 import 'components/input/character-check.dom';
 import 'components/cookies-banner/cookies-banner.dom';
 import 'components/button/button.dom';
