+++ src/patterns/addresses/_macro-options.md
